requirements.txt
ipython>=7.0.0
ipywidgets>=7.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
requests>=2.31.0
pydantic>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0 